% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{create_list_object}
\alias{create_list_object}
\title{Create a List Object with Custom Class}
\usage{
create_list_object(class, ...)
}
\arguments{
\item{class}{Character vector specifying the class(es) to assign to the object.}

\item{...}{Named elements to include in the list.}
}
\value{
A list with the specified class attribute.
}
\description{
Creates a structured list object with a specified class attribute.
}
\examples{
obj <- create_list_object(c("myclass", "list"), a = 1, b = 2)
class(obj)

}
