% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{odds_to_prob}
\alias{odds_to_prob}
\title{Convert Odds to Probability}
\usage{
odds_to_prob(odds)
}
\arguments{
\item{odds}{odds to convert to probability}
}
\value{
numeric probability value between 0 and 1
}
\description{
Converts odds to a probability value.
}
