% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{prob_to_odds}
\alias{prob_to_odds}
\title{Convert Probability to Odds}
\usage{
prob_to_odds(p)
}
\arguments{
\item{p}{a probability to convert to odds}
}
\value{
numeric odds value
}
\description{
Converts a probability value to odds.
}
