% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_input.r
\name{updateFormulaInput}
\alias{updateFormulaInput}
\title{Update Formula Input}
\usage{
updateFormulaInput(session, inputId, value = NULL, terms = NULL)
}
\arguments{
\item{session}{The session object passed to the Shiny server function.}

\item{inputId}{The id of the input object.}

\item{value}{The new value for the input.}

\item{terms}{A named list of terms to highlight. See \code{\link{formulaInput}}
for details on the format. Set to an empty list \code{list()} to clear
all term highlighting.}
}
\description{
Change the value and/or terms of a formula input on the client.
}
